outsourced version control stuff to own module
